refactor: remove kubehost option
